ckEditor: moved under components
