+++ webapp/components/ckEditor/ckEditorConfig.js
