+++ webapp/components/ckEditor/style.less
@@ -1,4 +1,4 @@
-@import (reference) "../app-styles/config";
+@import (reference) "../../app-styles/config";
 
 .cke_wrapper {
   .cke_top {
